+++ dart/lib/sentry_private.dart
@@ -1,0 +1,3 @@
+// attachments
+// ignore: invalid_export_of_internal_element
+export 'src/sentry_client_attachment_processor.dart';